box migrated to functional Component, Pnending optimisation
